Update README.md
You will write 10 Racket functions (not counting helper functions) and 1 Racket macro.

Download the starter code from the course website which includes some image files to make testing more fun, and the two Racket fileshw4.rkt and hw4tests.rkt. Add to these files to complete your homework.

Provided Code:

The code at the top of hw4tests.rkt uses a graphics library to provide a simple, entertaining (?) outlet for your streams. You need not understand this code (though it is not complicated) or even use it, but it may make the homework more fun. This is how you use it:

(open-window) returns a graphics window you can pass as the first argument to place-repeatedly.

(place-repeatedly window pause stream n) uses the first n values produced by stream. Each stream element must be a pair where the first value is an integer between 0 and 5 inclusive and the second value is a string that is the name of an image file (e.g., .jpg). (Sample image files that will work well are available on the course website. Put them in the same directory as your code.) Every pause seconds (where pause is a decimal, i.e., floating-point, number), the next stream value is retrieved, the corresponding image file is opened, and it is placed in the window using the number in the pair to choose its position in a 2×3 grid as follows: